backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
deep-translator>=1.11.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
